+++ src/connector.ts
@@ -16,6 +16,7 @@ import { disabledPrettyPrinter } from './printer';
 import { keywordize } from './util/string';
 import { initializeDebugger } from './debugger/calva-debug';
 import * as outputWindow from './results-output/results-doc';
+import { formatAsLineComments } from './results-output/util';
 import { evaluateInOutputWindow } from './evaluate';
 import * as liveShareSupport from './live-share';
 import * as calvaDebug from './debugger/calva-debug';
@@ -103,7 +104,9 @@ async function connectToHost(hostname: string, port: number, connectSequence: Re
     setStateValue('clj', cljSession);
     setStateValue('cljc', cljSession);
     status.update();
-    outputWindow.appendLine(`; Connected session: clj\n${outputWindow.CLJ_CONNECT_GREETINGS}`);
+    outputWindow.appendLine(
+      `; Connected session: clj\n${formatAsLineComments(outputWindow.CLJ_CONNECT_GREETINGS)}`
+    );
     replSession.updateReplSessionType();
 
     initializeDebugger(cljSession);
@@ -179,9 +182,9 @@ function setUpCljsRepl(session, build) {
   setStateValue('cljs', session);
   status.update();
   outputWindow.appendLine(
-    `; Connected session: cljs${build ? ', repl: ' + build : ''}\n${
+    `; Connected session: cljs${build ? ', repl: ' + build : ''}\n${formatAsLineComments(
       outputWindow.CLJS_CONNECT_GREETINGS
-    }`
+    )}`
   );
   outputWindow.setSession(session, 'cljs.user');
   replSession.updateReplSessionType();
@@ -630,6 +633,21 @@ export async function connect(
     }
   } catch (e) {
     console.error(e);
+  }
+  if (!['babashka', 'nbb', 'joyride', 'generic'].includes(connectSequence.projectType)) {
+    await nClient.session.info('clojure.core', 'map').catch((e) => {
+      void vscode.window
+        .showWarningMessage(
+          "Calva failed to perform a basic nREPL 'info' call. You need to start the REPL with cider-nrepl dependencies met",
+          'Show Calva Connect Docs'
+        )
+        .then((choice) => {
+          if (choice === 'Show Calva Connect Docs') {
+            void vscode.commands.executeCommand('simpleBrowser.show', 'https://calva.io/connect/');
+          }
+        });
+      console.error(`cider-nrepl dependencies not met: `, e);
+    });
   }
   return true;
 }