+++ test-data/projects/frankenstein/src/foo.clj
@@ -1,0 +1,3 @@
+(ns foo)
+
+(defn abs [])