+++ CHANGELOG.md
@@ -4,8 +4,10 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- Fix: [Warnings printed with an added `;` on a line of its own](https://github.com/BetterThanTomorrow/calva/issues/1930)
 - Fix: [Calva does not fall back on lsp definitions when nrepl definitions fail](https://github.com/BetterThanTomorrow/calva/issues/1933)
 - [Inform if Calva nREPL dependencies are not fulfilled](https://github.com/BetterThanTomorrow/calva/issues/1935)
+
 
 ## [2.0.314] - 2022-11-01